move AvatarGroup to Components folder.
