+++ src/app/(main)/my-voyage/tech-stack/components/TechStackCard.tsx
@@ -1,7 +1,7 @@
 import { PlusCircleIcon } from "@heroicons/react/24/outline";
 import type { TechItem } from "./fixtures/TechStack";
 import myAvatar from "@/public/img/avatar.png";
-import AvatarGroup from "./AvatarGroup";
+import AvatarGroup from "../../../../../components/AvatarGroup";
 import Avatar from "@/components/Avatar";
 import Button from "@/components/Button";
